hrms_backend/requirements.txt
asgiref==3.11.1
dj-database-url==3.1.2
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.16.1
gunicorn==25.1.0
mysqlclient==2.2.8
packaging==26.0
sqlparse==0.5.5
tzdata==2025.3
